+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## [Unreleased]
+* create build-time env-encryption key map to handle multiple possible env variables. 1 env == 1 key.
 
 ## 0.3.2
 * fixes key print output on windows